mcp/pyproject.toml
[project]
name = "connector-mcp"
version = "0.1.0"
description = "MCP server that exposes Airbyte Connector SDK as MCP tools"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Airbyte Support", email = "support@airbyte.io" }]

dependencies = [
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.0.0",
    "segment-analytics-python>=2.2.0",
    "jsonref>=1.1.0",
    "jsonpath-ng>=1.6.1",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "fastmcp>=2.10.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3,<9.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "ruff==0.7.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["connector_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
